feat: Implement CustomBottomNavigationBar for improved modularity and code reuse
